backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pillow==10.1.0
transformers==4.35.2
torch==2.1.1
torchvision==0.16.1
python-dotenv==1.0.0
pydantic==2.5.2
numpy==1.26.2
opencv-python==4.8.1.78 